refactor: MemoryAction 컴포넌트 내에서 이동되도록 수정
